+++ UniVgo2/Runtime/Enums/RenderPipelineType.cs
@@ -1,0 +1,17 @@
+﻿// ----------------------------------------------------------------------
+// @Namespace : UniVgo2
+// @Enum      : RenderPipelineType
+// ----------------------------------------------------------------------
+namespace UniVgo2
+{
+    /// <summary></summary>
+    public enum RenderPipelineType
+    {
+        /// <summary>Built-in Render Pipeline</summary>
+        BRP,
+        /// <summary>Universal Render Pipeline</summary>
+        URP,
+        /// <summary>High Definition Render Pipeline</summary>
+        HDRP,
+    }
+}